[TP-SPRINT #2.3] 마이페이지 - 내 정보: 내 계정 정보 관리 (#76)
* CBT routes 정리 / 계정 관리 1차 작성

* 내 정보 관리 - 유저 정보 변경 및 확인, 플랫폼 연동 관리 마무리

* 비밀번호 확인 -> 비밀번호 틀림 알림 문구 추가

* 프로필사진 기능 적용 / 기능제안 author 수정 ( string->userEntity

* 기능제안 작성자 truepoint 고정, entity 변경에 따른 관리자 오류 수정

* 회원탈퇴 기능 추가

* 앱바 유저 메뉴 "내정보 변경" 올바르게 링크

* 닉네임 입력 -> 선택 방식 변경 / 이미 채널 연동된 경우 처리

* 다이얼로그 스크롤락 제거

* 플랫폼 연동 스낵바 가독성 수정, 플랫폼 연동 다이얼로그 연동 플랫폼 없는 경우 배치 수정

* 프로필 이미지 연동 제거한 뒤에도 선택할 수 있는 버그 수정, 유튜브 연동 제거시 유튜브 대표프사 제거 되지 않는 버그 수정

* merge conflict 수정

* 프사/닉네임 사용안함 추가 + 리렌더링에 새로운 데이터 요청 추가

+++ server/src/resources/users/users.service.ts
@@ -1,9 +1,13 @@
 import bcrypt from 'bcrypt';
 import {
-  Injectable, HttpException, HttpStatus, BadRequestException,
+  Injectable, HttpException, HttpStatus, BadRequestException, InternalServerErrorException,
 } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
+import { UpdateUserDto } from '@truepoint/shared/dist/dto/users/updateUser.dto';
+import { ProfileImages } from '@truepoint/shared/dist/res/ProfileImages.interface';
+import { ChannelNames } from '@truepoint/shared/dist/res/ChannelNames.interface';
+import { LinkPlatformRes } from '@truepoint/shared/dist/res/LinkPlatformRes.interface';
 import { UserEntity } from './entities/user.entity';
 import { UserTokenEntity } from './entities/userToken.entity';
 import { SubscribeEntity } from './entities/subscribe.entity';
@@ -28,6 +32,10 @@ export class UsersService {
     private readonly youtubeRepository: Repository<PlatformYoutubeEntity>,
   ) {}
 
+  private resizeingYoutubeLogo(youtubeLogoString: string): string {
+    return youtubeLogoString.replace('{size}', '150');
+  }
+
   async findAll(): Promise<UserEntity[]> {
     return this.usersRepository.find();
   }
@@ -43,8 +51,89 @@ export class UsersService {
     });
   }
 
-  async remove(userid: string): Promise<void> {
-    await this.usersRepository.delete(userid);
+  /**
+   * 특정 유저의 연동된 플랫폼의 프로필 이미지 정보를 반환하는 메서드
+   * @param userId 프로필 이미지 정보를 열람하고자 하는 유저 아이디
+   */
+  async findOneProfileImage(userId: string): Promise<ProfileImages> {
+    const user = await this.usersRepository.findOne(userId);
+    const images = [];
+
+    // 아프리카는 OPEN API 업데이트 이후 추가 20.11.18 hwasurr
+    // const afreecaLink = await this.afreecaRepository.findOne(user.afreecaId);
+    // if (afreecaLink) images.push({ platform: 'afreeca', logo: afreecaLink.logo });
+    if (user.twitchId) {
+      const twitchLink = await this.twitchRepository.findOne(user.twitchId);
+      if (twitchLink) images.push({ platform: 'twitch', logo: twitchLink.logo });
+    }
+
+    if (user.youtubeId) {
+      const youtubeLink = await this.youtubeRepository.findOne(user.youtubeId);
+      if (youtubeLink) {
+        images.push(
+          { platform: 'youtube', logo: this.resizeingYoutubeLogo(youtubeLink.youtubeLogo) },
+        );
+      }
+    }
+    return images;
+  }
+
+  async findChannelNames(userId: string): Promise<ChannelNames> {
+    const user = await this.usersRepository.findOne(userId);
+    const nickNames: ChannelNames = [];
+
+    // 아프리카는 OPEN API 업데이트 이후 추가 20.11.18 hwasurr
+    // const afreecaLink = await this.afreecaRepository.findOne(user.afreecaId);
+    // if (afreecaLink) images.push({ platform: 'afreeca', logo: afreecaLink.logo });
+
+    if (user.twitchId) {
+      const twitchLink = await this.twitchRepository.findOne(user.twitchId);
+      if (twitchLink) nickNames.push({ platform: 'twitch', nickName: twitchLink.twitchChannelName });
+    }
+
+    if (user.youtubeId) {
+      const youtubeLink = await this.youtubeRepository.findOne(user.youtubeId);
+      if (youtubeLink) {
+        nickNames.push(
+          { platform: 'youtube', nickName: youtubeLink.youtubeTitle },
+        );
+      }
+    }
+    return nickNames;
+  }
+
+  async updateOne(dto: UpdateUserDto): Promise<any> {
+    return this.usersRepository.save({
+      userId: dto.userId,
+      gender: dto.gender,
+      mail: dto.mail,
+      nickName: dto.nickName,
+      profileImage: dto.profileImage,
+    });
+  }
+
+  /**
+   * 유저 정보 삭제 메서드 userId를 제외한 모든 column을 없앱니다.
+   * @param userId 삭제할 유저 고유 아이디
+   */
+  async remove(userId: string): Promise<number> {
+    const targetUser = await this.usersRepository.findOne(userId);
+    const result = await this.usersRepository.update(userId, {
+      name: '',
+      mail: '',
+      nickName: '',
+      userDI: `DELETED_${targetUser.userDI}`,
+      profileImage: '',
+      afreecaId: '',
+      youtubeId: '',
+      twitchId: '',
+      password: '',
+      birth: '',
+      gender: '',
+      marketingAgreement: false,
+      phone: '',
+    });
+    return result.affected;
   }
 
   async register(user: UserEntity): Promise<UserEntity> {
@@ -91,12 +180,11 @@ export class UsersService {
   }
 
   // 본인인증의 결과가 인증이 되면,  해당 계정의 패스워드를 변경한다.
-  async findPW(userDI: string, password: string): Promise<boolean> {
+  async updatePW(userDI: string, password: string): Promise<boolean> {
     try {
       const user = await this.usersRepository
         .findOne({ where: { userDI } });
       if (user) {
-        // 임시번호 저장 후에 임시비밀번호 저장.
         const hashedPassword = await bcrypt.hash(password, 10);
 
         await this.usersRepository
@@ -111,7 +199,7 @@ export class UsersService {
       }
       return false;
     } catch {
-      throw new HttpException('findPW error', HttpStatus.BAD_REQUEST);
+      throw new HttpException('updatePW error', HttpStatus.INTERNAL_SERVER_ERROR);
     }
   }
 
@@ -188,6 +276,9 @@ export class UsersService {
   // Refresh Token 삭제 - 로그아웃을 위해
   async removeOneToken(userId: string): Promise<UserTokenEntity> {
     const userToken = await this.userTokensRepository.findOne(userId);
+    if (!userToken) {
+      throw new InternalServerErrorException('userToken waht you request to logout is not exists');
+    }
     return this.userTokensRepository.remove(userToken);
   }
 
@@ -203,8 +294,8 @@ export class UsersService {
 
   async linkUserToPlatform(
     userId: string, platform: string, platformId: string,
-  ): Promise<UserEntity | null> {
-    // 적재된 연동 데이터의 실제 트루포인트 유저 정보에 추가
+  ): Promise<LinkPlatformRes> {
+    // 적재된 연동 데이터의 고유 아이디를 실제 트루포인트 유저 정보에 추가
     // 어느 트루포인트 유저가 트위치 연동을 진행했는 지 곧바로 알 수 없음..
     // twitch accesstoken 발급 이후 프론트로 리다이렉트할 때 방금 유저의 고유 id와 플랫폼 정보를 전달하고,
     // 그 정보를 다시 백엔드에서 받아와 올바르 유저에 맞게 적재하는 방식.
@@ -215,18 +306,39 @@ export class UsersService {
     // platformTwitch, platformYoutube, platformAfreeca 등을 체크한 후 있으면 그 때 넣도록 처리.
     switch (platform) {
       case 'twitch': {
+        // 이미 적재된 경우 다시 적재하지 않음.
+        if (targetUser.twitchId === platformId) return 'already-linked';
         const linkedInfo = await this.twitchRepository.findOne(platformId);
         if (linkedInfo) targetUser[`${platform}Id`] = platformId;
+        else {
+          throw new InternalServerErrorException(
+            'An error occurred during linking platform: there is no platform information',
+          );
+        }
         break;
       }
       case 'youtube': {
+        // 이미 적재된 경우 다시 적재하지 않음.
+        if (targetUser.youtubeId === platformId) return 'already-linked';
         const linkedInfo = await this.youtubeRepository.findOne(platformId);
         if (linkedInfo) targetUser[`${platform}Id`] = platformId;
+        else {
+          throw new InternalServerErrorException(
+            'An error occurred during linking platform: there is no platform information',
+          );
+        }
         break;
       }
       case 'afreeca': {
+        // 이미 적재된 경우 다시 적재하지 않음.
+        if (targetUser.afreecaId === platformId) return 'already-linked';
         const linkedInfo = await this.afreecaRepository.findOne(platformId);
         if (linkedInfo) targetUser[`${platform}Id`] = platformId;
+        else {
+          throw new InternalServerErrorException(
+            'An error occurred during linking platform: there is no platform information',
+          );
+        }
         break;
       }
       default: throw new BadRequestException('platform must be one of "twitch" | "afreeca" | "youtube"');
@@ -259,13 +371,36 @@ export class UsersService {
 
   /**
    * 연결된 트루포인트 Users 에서 플랫폼 연결 link를 삭제하는 메소드
+   * 부가적으로, 대표 프로필 사진이 삭제하고자하는 플랫폼으로 설정되어 있는 경우 대표 프로필 사진을 삭제합니다.
    * @param {string} userId 트루포인트 유저 아이디 문자열
    * @param {string} platform 연결된 플랫폼 문자열
    */
   async disconnectLink(
     userId: string, platform: string,
   ): Promise<number> {
-    const result = await this.usersRepository.update(userId, { [`${platform}Id`]: undefined });
+    const targetUser = await this.usersRepository.findOne(userId);
+    let targetPlatformLogo: string;
+    // 아프리카의 경우 아직 채널/유저 프로필사진 데이터를 가져올 수 없다.
+    // if (platform === 'afreeca') {
+    //   const afreeca = await this.twitchRepository.findOne(targetUser.afreecaId);
+    //   targetPlatformLogo = afreeca.logo;
+    // }
+    if (platform === 'twitch') {
+      const twitch = await this.twitchRepository.findOne(targetUser.twitchId);
+      targetPlatformLogo = twitch.logo;
+    }
+    if (platform === 'youtube') {
+      const youtube = await this.youtubeRepository.findOne(targetUser.youtubeId);
+      targetPlatformLogo = this.resizeingYoutubeLogo(youtube.youtubeLogo);
+    }
+
+    // 플랫폼 연결 정보 삭제 및 대표 프로필 사진이 해당 플랫폼의 프로필사진인 경우 함께 삭제
+    const result = await this.usersRepository.update(
+      userId, {
+        [`${platform}Id`]: undefined,
+        profileImage: targetPlatformLogo === targetUser.profileImage ? undefined : targetUser.profileImage,
+      },
+    );
     return result.affected;
   }
